Bump dependencies in poetry
During the process of developing this work there were two dependency
managers used:

  - pipenv
  - poetry

The tests, production and other build objects are executed with the
"poetry" dependency management system.

Because the development occurred against the pipenv version,
dependencies were not committed to the pipenv dependency management
tooling.

This commit addresses that by adding the required dependencies, allowing
the E2E tests to execute.