Default an empty object for the spec query
Within the links function there is a query for the "spec" part of a
given object. However, spec is not part of the "standard object design"
of Kubernetes; in some cases that object may be missing.

This commit addresses that by querying the object regardless, but
defaulting to an empty object in the case that property is missing.